working program basic algorithm
done: working program, basic algorithm, basic random method for generating valid sudokus
todo: rethink oop design, maybe add patterns for acomodating future algorithms
